fix: harden I2I support — unify NvExt, fix truthy fallbacks, align seed type

+++ lib/llm/src/protocols/openai/images.rs
@@ -11,18 +11,26 @@ mod nvext;
 pub use aggregator::DeltaAggregator;
 pub use nvext::{NvExt, NvExtProvider};
 
+/// Image generation request with NVIDIA extensions.
+///
+/// Generation params (seed, guidance_scale, etc.) can be specified at the top level
+/// (SGLang-native format) or nested under `nvext` (Dynamo convention). The Python
+/// handler merges them, with `nvext` taking precedence. Top-level fields exist because
+/// SGLang's native API uses them, and requiring `nvext` wrapping would break SGLang
+/// client compatibility.
 #[derive(Serialize, Deserialize, Validate, Debug, Clone)]
 pub struct NvCreateImageRequest {
     #[serde(flatten)]
     pub inner: dynamo_async_openai::types::CreateImageRequest,
 
-    /// Optional image reference that guides generation (for I2I)
+    /// Optional image reference that guides generation (for I2I/TI2I).
     #[serde(skip_serializing_if = "Option::is_none")]
     pub input_reference: Option<String>,
 
     /// Random seed. Top-level alternative to nvext; nvext takes precedence.
+    /// i64 to match PyTorch's torch.manual_seed() accepted range.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub seed: Option<u32>,
+    pub seed: Option<i64>,
 
     /// Negative prompt. Top-level alternative to nvext; nvext takes precedence.
     #[serde(skip_serializing_if = "Option::is_none")]
